# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Doc",
        "version": "1.0"
    },
    "paths": {
...
        "/api/posts": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "Get all posts",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "posts": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "title": {
                                                                "type": "string",
                                                                "example": "Some title"
                                                            },
                                                            "content": {
                                                                "type": "string",
                                                                "example": "Some content"
                                                            },
                                                            "created_at": {
                                                                "type": "timestamp",
                                                                "example": "2023-06-17T18:31:55.000000Z"
                                                            },
                                                            "updated_at": {
                                                                "type": "timestamp",
                                                                "example": "2023-06-17T18:31:55.000000Z"
                                                            },
                                                            "image": {
                                                                "type": "string",
                                                                "example": "public/images/image1.png"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "authToken": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Post"
                ],
                "summary": "Create post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "title": {
                                                "type": "string",
                                                "example": "Some title"
                                            },
                                            "content": {
                                                "type": "string",
                                                "example": "Some content"
                                            },
                                            "image": {
                                                "type": "image",
                                                "example": "image1.png"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "authToken": []
                    }
                ]
            }
        },
        "/api/posts/{post}": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "Get one post",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "Post ID",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "post": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Some title"
                                                },
                                                "content": {
                                                    "type": "string",
                                                    "example": "Some content"
                                                },
                                                "created_at": {
                                                    "type": "timestamp",
                                                    "example": "2023-06-17T18:31:55.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "timestamp",
                                                    "example": "2023-06-17T18:31:55.000000Z"
                                                },
                                                "image": {
                                                    "type": "string",
                                                    "example": "public/images/image1.png"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "authToken": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Post"
                ],
                "summary": "Delete post",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "Post ID",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "authToken": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Post"
                ],
                "summary": "Update post",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "Post ID",
                        "required": true,
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Some title"
                                            },
                                            "content": {
                                                "type": "string",
                                                "example": "Some content"
                                            },
                                            "image": {
                                                "type": "image",
                                                "example": "image1.png"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "authToken": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "authToken": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}